1) Fixed issues with resizing on the same node
Previously container would be stopped and "transferred" to itself,
resulting in a conflict when lxc sees the same profile name. Now
it checks if the destination node is the same as the source, and
just updates the lxc profile without stopping the container.

2) Fixed issue with confirm_migration(): 1) taking incorrect number of
parameters, and 2) skips unplugging of vifs if source and destination
nodes are the same.

Cherry-picked from: ca986d9265b0cfe49d04ebb308986f9d0dd1e9aa